move merge logic into flags extension
